pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gpt_pptx_summarizer"
version = "0.1"
description = "A package for summarizing PPTX files using GPT"
readme = "README.md"
requires-python = ">=3.6"
dependencies = [
    "flask",
    "requests",
    "python-pptx",
    "openai",
    "pytest",
    "pytest-asyncio"
]

[project.urls]
homepage = "https://github.com/Scaleup-Excellenteam/final-exercise-DSH93"
